Renamed cpairs to pair_counting_engines
