+++ halotools/mock_observables/pair_counters/npairs_jackknife_3d.py
@@ -11,7 +11,7 @@ __author__ = ('Andrew Hearin', 'Duncan Campbell')
 
 from .rectangular_mesh import RectangularDoubleMesh
 from .mesh_helpers import _set_approximate_cell_sizes, _cell1_parallelization_indices
-from .cpairs import npairs_jackknife_3d_engine
+from .pair_counting_engines import npairs_jackknife_3d_engine
 from .npairs_3d import _npairs_3d_process_args
 
 from ...custom_exceptions import HalotoolsError
