+++ halotools/mock_observables/pair_counters/npairs_per_object_3d.py
@@ -10,7 +10,7 @@ __author__ = ('Andrew Hearin', 'Duncan Campbell')
 
 from .rectangular_mesh import RectangularDoubleMesh
 from .mesh_helpers import _set_approximate_cell_sizes, _cell1_parallelization_indices
-from .cpairs import npairs_per_object_3d_engine
+from .pair_counting_engines import npairs_per_object_3d_engine
 from .npairs_3d import _npairs_3d_process_args
 
 __all__ = ('npairs_per_object_3d', )
