+++ halotools/mock_observables/pair_counters/pair_counting_engines/npairs_3d_engine.pyx
