+++ halotools/mock_observables/pair_counters/pair_counting_engines/npairs_jackknife_3d_engine.pyx
