+++ halotools/mock_observables/pair_counters/pair_counting_engines/npairs_s_mu_engine.pyx
