+++ halotools/mock_observables/pair_counters/pairwise_distance_3d.py
@@ -10,7 +10,7 @@ __author__ = ('Andrew Hearin', 'Duncan Campbell')
 
 from .rectangular_mesh import RectangularDoubleMesh
 from .mesh_helpers import _set_approximate_cell_sizes, _enclose_in_box, _cell1_parallelization_indices
-from .cpairs import pairwise_distance_3d_engine
+from .pair_counting_engines import pairwise_distance_3d_engine
 from ...utils.array_utils import convert_to_ndarray, array_is_monotonic, custom_len
 from scipy.sparse import coo_matrix
